Upgrade stylus-loader
Running dev from a clean install resulted in webpack throwing a mountain
of errors that all said "Path must be a string. Received undefined". It
looks like stylus-loader removed the explicit dependency on stylus in a
minor release. Updating stylus-loader and depending on stylus explicitly
appears to have fixed the issue.